AbsScan is now supporting the case where the solutions vector is empty. It then does a simple bounday scan with linear interpolation. This is the scan which before was residing in MethodDatasetsPluginScan. I have moved it from there.
