+++ core/include/MethodAbsScan.h
@@ -18,6 +18,7 @@
 #include "TPaveText.h"
 #include "TF1.h"
 #include "TDatime.h"
+#include "TMVA/TSpline1.h"
 
 #include "Utils.h"
 #include "OneMinusClPlotAbs.h"
@@ -119,6 +120,8 @@ class MethodAbsScan
 		inline void                     setVerbose(bool yesNo=true){verbose = yesNo;};
 		void 							setXscanRange(float min, float max);
 		void 							setYscanRange(float min, float max);
+		void							calcCLintervalsSimple();
+		const std::pair<double, double> getBorders(const TGraph& graph, const double confidence_level);
 
 		vector<RooSlimFitResult*> allResults;           ///< All fit results we encounter along the scan.
 		vector<RooSlimFitResult*> curveResults;         ///< All fit results of the the points that make it into the 1-CL curve.
