+++ core/src/MethodDatasetsPluginScan.cpp
@@ -11,7 +11,6 @@
 
 #include "MethodDatasetsPluginScan.h"
 #include "TRandom3.h"
-#include "TMVA/TSpline1.h"
 #include <algorithm>
 #include <ios>
 #include <iomanip>
@@ -1366,53 +1365,3 @@ void MethodDatasetsPluginScan::printDebug(const RooFitResult& r){
           << std::resetiosflags(std::ios::fixed) 
           << std::resetiosflags(std::ios::scientific);
 };
-
-void MethodDatasetsPluginScan::calcCLintervals()
-{
- 
-  clintervals1sigma.clear(); 
-  clintervals2sigma.clear();
-  
-  double levels[2] = {0.6827, 0.9545};
-  for (int c=0;c<2;c++){
-    const std::pair<double, double> borders = getBorders(TGraph(this->hCL), levels[c]);
-    CLInterval cli;
-    cli.pvalue = 1. - levels[c];
-    cli.min = borders.first;
-    cli.max = borders.second;
-    cli.central = -1;
-    if ( c==0 ) clintervals1sigma.push_back(cli);
-    if ( c==1 ) clintervals2sigma.push_back(cli);
-    std::cout<<"borders at "<<levels[c]<<"  [ "<<borders.first<<" : "<<borders.second<<"]"<<std::endl;
-  }
-
-}
-
-const std::pair<double, double> MethodDatasetsPluginScan::getBorders(const TGraph& graph, const double confidence_level){
-
-  const double p_val = 1 - confidence_level;
-  // performing a conservative linear interpolation
-  TMVA::TSpline1 splines("splines", new TGraph(graph));
-
-  double min_edge = graph.GetX()[0];
-  // will never return smaller edge than min_edge
-  double max_edge = graph.GetX()[graph.GetN()-1];
-  // will never return higher edge than max_edge
-  int scan_steps = 1000;
-  double lower_edge = min_edge;
-  double upper_edge = max_edge;
-
-  for(double point = min_edge; point < max_edge; point+= (max_edge-min_edge)/scan_steps){
-    if(splines.Eval(point)>p_val){
-      lower_edge = point;
-      break;
-    }
-  }
-  for(double point = max_edge; point > min_edge; point-= (max_edge-min_edge)/scan_steps){
-    if(splines.Eval(point)>p_val){
-      upper_edge = point;
-      break;
-    }
-  }
-  return std::pair<double, double>(lower_edge,upper_edge);
-}
